backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9

pymupdf==1.24.9
pillow==10.4.0
numpy==1.26.4

faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
torch>=2.2.0
transformers==4.44.2

paddleocr==2.8.1
paddlepaddle==2.6.2

rapidfuzz==3.9.6
tqdm==4.66.5

requests==2.32.3

# Mathematical expression formatting
sympy==1.14.0
rich==13.7.1
